+++ requirements.txt
@@ -1,8 +1,9 @@
 PyQt5 >= 5.13
 python-vlc
 PyQtGraph
-pyechart
+PyQtChart
 influxdb
 opencv-python
 pandas
 psutil
+scipy
