+++ src/cal_ext_coef.py
@@ -4,6 +4,7 @@ import os
 import numpy as np
 import pandas as pd
 import traceback
+import warnings
 
 import scipy
 from scipy.optimize import curve_fit
@@ -35,6 +36,7 @@ def select_max_rgb(r, g, b):
 def cal_curve(hanhwa: pd.DataFrame):
     # hanhwa = pd.read_csv(f"{rgbsavedir}/{epoch}.csv")
     # print(hanhwa)
+    global hanhwa_opt_r, hanhwa_opt_g, hanhwa_opt_b
     hanhwa = hanhwa.sort_values(by=['distance'])
     hanhwa_dist = hanhwa[['distance']].squeeze().to_numpy()
     hanhwa_x = np.linspace(hanhwa_dist[0], hanhwa_dist[-1], 100, endpoint=True)
@@ -58,14 +60,21 @@ def cal_curve(hanhwa: pd.DataFrame):
     b_ext_init = [b1_init, b2_init, 1]
 
     try:
-
         hanhwa_opt_r, hanhwa_cov_r = curve_fit(func, hanhwa_dist, hanhwa_r, p0=r_ext_init, maxfev=5000)
         hanhwa_opt_g, hanhwa_cov_g = curve_fit(func, hanhwa_dist, hanhwa_g, p0=g_ext_init, maxfev=5000)
         hanhwa_opt_b, hanhwa_cov_b = curve_fit(func, hanhwa_dist, hanhwa_b, p0=b_ext_init, maxfev=5000)
 
-    except Exception:
-        print(traceback.format_exc())
-        return
+        # hanhwa_opt_r, hanhwa_cov_r = curve_fit(func, hanhwa_dist, hanhwa_r, p0=r_ext_init)
+        # hanhwa_opt_g, hanhwa_cov_g = curve_fit(func, hanhwa_dist, hanhwa_g, p0=g_ext_init)
+        # hanhwa_opt_b, hanhwa_cov_b = curve_fit(func, hanhwa_dist, hanhwa_b, p0=b_ext_init)
+
+    except RuntimeError:
+        # Optimal parameters not found: Number of calls to function has reached maxfev = 5000.
+        # 이 에러는 타겟의 RGB에 대해 최적의 기울기를 찾지 못하여 발생하는 에러로,
+        # 타겟이 이상할 때 대게 발생함.
+
+        # print(traceback.format_exc())
+        pass
 
     list1 = []
     list2 = []
@@ -100,23 +109,24 @@ def cal_curve(hanhwa: pd.DataFrame):
 
 # @staticmethod
 def func(x, c1, c2, a):
+    warnings.filterwarnings('ignore')
     return c2 + (c1 - c2) * np.exp(-a * x)
 
 
-def print_result(opt_r, opt_g, opt_b, err_r, err_g, err_b):
-    print(f'Red channel: (',
-          f'C1: {opt_r[0]:.2f} ± {err_r[0]:.2f}, ',
-          f'C2: {opt_r[1]:.2f} ± {err_r[1]:.2f}, ',
-          f'alpha: {opt_r[2]:.2f} ± {err_r[2]:.2f})')
-    print(f'Green channel: (',
-          f'C1: {opt_g[0]:.2f} ± {err_g[0]:.2f}, ',
-          f'C2: {opt_g[1]:.2f} ± {err_g[1]:.2f}, ',
-          f'alpha: {opt_g[2]:.2f} ± {err_g[2]:.2f})')
-    print(f'Blue channel: (',
-          f'C1: {opt_b[0]:.2f} ± {err_b[0]:.2f}, ',
-          f'C2: {opt_b[1]:.2f} ± {err_b[1]:.2f}, ',
-          f'alpha: {opt_b[2]:.2f} ± {err_b[2]:.2f})')
+# def print_result(opt_r, opt_g, opt_b, err_r, err_g, err_b):
+#     print(f'Red channel: (',
+#           f'C1: {opt_r[0]:.2f} ± {err_r[0]:.2f}, ',
+#           f'C2: {opt_r[1]:.2f} ± {err_r[1]:.2f}, ',
+#           f'alpha: {opt_r[2]:.2f} ± {err_r[2]:.2f})')
+#     print(f'Green channel: (',
+#           f'C1: {opt_g[0]:.2f} ± {err_g[0]:.2f}, ',
+#           f'C2: {opt_g[1]:.2f} ± {err_g[1]:.2f}, ',
+#           f'alpha: {opt_g[2]:.2f} ± {err_g[2]:.2f})')
+#     print(f'Blue channel: (',
+#           f'C1: {opt_b[0]:.2f} ± {err_b[0]:.2f}, ',
+#           f'C2: {opt_b[1]:.2f} ± {err_b[1]:.2f}, ',
+#           f'alpha: {opt_b[2]:.2f} ± {err_b[2]:.2f})')
 
 
-def extcoeff_to_vis(optimal, error, coeff=3.912):
-    return coeff / (optimal + np.array((1, 0, -1)) * error)
+# def extcoeff_to_vis(optimal, error, coeff=3.912):
+#     return coeff / (optimal + np.array((1, 0, -1)) * error)
