+++ src/resources/clock.png
