+++ src/resources/setting_window.ui
@@ -6,8 +6,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>1159</width>
-    <height>888</height>
+    <width>971</width>
+    <height>558</height>
    </rect>
   </property>
   <property name="sizePolicy">
@@ -108,7 +108,7 @@ color: #ffffff;</string>
                 </property>
                 <property name="icon">
                  <iconset>
-                  <normaloff>asset/flip_off.png</normaloff>asset/flip_off.png</iconset>
+                  <normaloff>flip_off.png</normaloff>flip_off.png</iconset>
                 </property>
                 <property name="iconSize">
                  <size>
