+++ src/resources/thumb.ui
@@ -9,8 +9,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>1035</width>
-    <height>816</height>
+    <width>784</width>
+    <height>366</height>
    </rect>
   </property>
   <property name="sizePolicy">
@@ -22,7 +22,7 @@
   <property name="minimumSize">
    <size>
     <width>400</width>
-    <height>300</height>
+    <height>100</height>
    </size>
   </property>
   <property name="font">
@@ -71,7 +71,7 @@ border-color: rgb(255, 255, 255);</string>
      <number>0</number>
     </property>
     <property name="topMargin">
-     <number>0</number>
+     <number>2</number>
     </property>
     <property name="rightMargin">
      <number>0</number>
@@ -85,13 +85,13 @@ border-color: rgb(255, 255, 255);</string>
     <item row="0" column="0">
      <layout class="QVBoxLayout" name="verticalLayout">
       <property name="spacing">
-       <number>2</number>
+       <number>0</number>
       </property>
       <property name="leftMargin">
        <number>0</number>
       </property>
       <property name="topMargin">
-       <number>2</number>
+       <number>0</number>
       </property>
       <property name="rightMargin">
        <number>0</number>
@@ -105,7 +105,7 @@ border-color: rgb(255, 255, 255);</string>
          <number>2</number>
         </property>
         <property name="sizeConstraint">
-         <enum>QLayout::SetDefaultConstraint</enum>
+         <enum>QLayout::SetMinimumSize</enum>
         </property>
         <item>
          <widget class="QLabel" name="front_label">
@@ -118,9 +118,7 @@ border-color: rgb(255, 255, 255);</string>
           <property name="font">
            <font>
             <family>Noto Sans</family>
-            <pointsize>17</pointsize>
-            <weight>50</weight>
-            <bold>false</bold>
+            <pointsize>15</pointsize>
            </font>
           </property>
           <property name="styleSheet">
@@ -146,9 +144,7 @@ color: rgb(255, 255, 255);</string>
           <property name="font">
            <font>
             <family>Noto Sans</family>
-            <pointsize>17</pointsize>
-            <weight>50</weight>
-            <bold>false</bold>
+            <pointsize>15</pointsize>
            </font>
           </property>
           <property name="styleSheet">
@@ -168,27 +164,35 @@ color: rgb(255, 255, 255);</string>
       <item>
        <layout class="QHBoxLayout" name="image_horizontalLayout">
         <property name="spacing">
-         <number>2</number>
+         <number>6</number>
+        </property>
+        <property name="sizeConstraint">
+         <enum>QLayout::SetMaximumSize</enum>
         </property>
         <item>
          <widget class="QLabel" name="front_image">
           <property name="sizePolicy">
-           <sizepolicy hsizetype="Preferred" vsizetype="Preferred">
+           <sizepolicy hsizetype="Preferred" vsizetype="Minimum">
             <horstretch>0</horstretch>
             <verstretch>0</verstretch>
            </sizepolicy>
           </property>
+          <property name="minimumSize">
+           <size>
+            <width>0</width>
+            <height>0</height>
+           </size>
+          </property>
           <property name="font">
            <font>
-            <family>Noto Sans</family>
-            <pointsize>20</pointsize>
+            <pointsize>30</pointsize>
            </font>
           </property>
           <property name="styleSheet">
-           <string notr="true">color: rgb(255, 255, 255);</string>
+           <string notr="true"/>
           </property>
           <property name="text">
-           <string>No file</string>
+           <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;br/&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
           </property>
           <property name="alignment">
            <set>Qt::AlignCenter</set>
@@ -198,22 +202,27 @@ color: rgb(255, 255, 255);</string>
         <item>
          <widget class="QLabel" name="rear_image">
           <property name="sizePolicy">
-           <sizepolicy hsizetype="Preferred" vsizetype="Preferred">
+           <sizepolicy hsizetype="Preferred" vsizetype="Minimum">
             <horstretch>0</horstretch>
             <verstretch>0</verstretch>
            </sizepolicy>
           </property>
+          <property name="minimumSize">
+           <size>
+            <width>0</width>
+            <height>0</height>
+           </size>
+          </property>
           <property name="font">
            <font>
-            <family>Noto Sans</family>
-            <pointsize>20</pointsize>
+            <pointsize>30</pointsize>
            </font>
           </property>
           <property name="styleSheet">
-           <string notr="true">color: rgb(255, 255, 255);</string>
+           <string notr="true"/>
           </property>
           <property name="text">
-           <string>No file</string>
+           <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;br/&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
           </property>
           <property name="alignment">
            <set>Qt::AlignCenter</set>
