+++ src/target_info.py
@@ -4,7 +4,7 @@ import pandas as pd
 import numpy as np
 
 import cal_ext_coef
-from model import JS06Settings
+from model import JS08Settings
 
 
 def minprint(epoch, left_range, right_range, distance, cv_img, camera):
@@ -21,6 +21,7 @@ def minprint(epoch, left_range, right_range, distance, cv_img, camera):
 
     visibility = get_rgb(epoch, min_x, min_y, cp_image, distance, camera)
 
+    # print(f'minprint visibility: {visibility}')
     return visibility
 
 
@@ -64,6 +65,7 @@ def get_rgb(epoch: str, min_x, min_y, cp_image, distance, camera):
 
     visibility = save_rgb(r_list, g_list, b_list, epoch, distance, camera)
 
+    # print(f'get_rgb visibility: {visibility}')
     return visibility
 
 
@@ -73,29 +75,35 @@ def save_rgb(r_list, g_list, b_list, epoch, distance, camera):
     save_path = os.path.join(f'rgb/{camera}')
     os.makedirs(save_path, exist_ok=True)
 
-    if r_list:
-        r_list = list(map(int, r_list))
-        g_list = list(map(int, g_list))
-        b_list = list(map(int, b_list))
+    try:
+        if r_list:
+            r_list = list(map(int, r_list))
+            g_list = list(map(int, g_list))
+            b_list = list(map(int, b_list))
 
-        col = ['target_name', 'r', 'g', 'b', 'distance']
-        result = pd.DataFrame(columns=col)
-        result['target_name'] = [f'target_{num}' for num in range(1, len(r_list) + 1)]
-        result['r'] = r_list
-        result['g'] = g_list
-        result['b'] = b_list
-        result['distance'] = distance
-        result.to_csv(f'{save_path}/{epoch}.csv', mode='w', index=False)
-        list1, list2, list3, select_color = cal_ext_coef.cal_curve(result)
+            col = ['target_name', 'r', 'g', 'b', 'distance']
+            result = pd.DataFrame(columns=col)
+            result['target_name'] = [f'target_{num}' for num in range(1, len(r_list) + 1)]
+            result['r'] = r_list
+            result['g'] = g_list
+            result['b'] = b_list
+            result['distance'] = distance
+            result.to_csv(f'{save_path}/{epoch}.csv', mode='w', index=False)
+            list1, list2, list3, select_color = cal_ext_coef.cal_curve(result)
 
-        visibility = extinc_print(list1, list2, list3, select_color)
+            # visibility = extinc_print(list1, list2, list3, select_color)
+            visibility = extinc_print(list3, select_color)
 
-    return visibility
+            # print(f'save_rgb visibility: {visibility}')
+            return visibility
+
+    except TypeError:
+        pass
 
 
-def extinc_print(c1_list: list = [0, 0, 0], c2_list: list = [0, 0, 0], alp_list: list = [0, 0, 0],
-                 select_color: str = ""):
+def extinc_print(alp_list: list, select_color: str = ""):
     """Select an appropriate value among visibility by wavelength."""
+    visibility = 0
 
     if select_color == 'red':
         visibility = visibility_print(alp_list[0])
@@ -104,6 +112,7 @@ def extinc_print(c1_list: list = [0, 0, 0], c2_list: list = [0, 0, 0], alp_list:
     elif select_color == 'blue':
         visibility = visibility_print(alp_list[2])
 
+    # print(f'extinc_print visibility: {visibility}')
     return visibility
 
 
@@ -124,21 +133,55 @@ def visibility_print(ext_g: float = 0.0):
 def get_target(camera_name: str):
     """Retrieves target information of a specific camera."""
 
-    save_path = JS06Settings.get('target_csv_path')
+    save_path = JS08Settings.get('target_csv_path')
 
     if os.path.isfile(f'{save_path}/{camera_name}/{camera_name}.csv'):
         target_df = pd.read_csv(f'{save_path}/{camera_name}/{camera_name}.csv')
         target_name = target_df['target_name'].tolist()
         left_range = str_to_tuple(target_df['left_range'].tolist())
-        # left_range = str_to_tuple(left_range)
         right_range = str_to_tuple(target_df['right_range'].tolist())
-        # right_range = str_to_tuple(right_range)
-        azimuth = target_df['azimuth'].tolist()
         distance = target_df['distance'].tolist()
+        azimuth = target_df['azimuth'].tolist()
+
         return target_name, left_range, right_range, distance, azimuth
 
     else:
         return [], [], [], [], []
+
+
+def get_target_from_azimuth(camera_name: str, azimuth: str):
+    """Retrieves target information from azimuth of a specific camera"""
+
+    global target_name, left_range, right_range, distance
+    save_path = JS08Settings.get('target_csv_path')
+
+    if os.path.isfile(f'{save_path}/{camera_name}/{camera_name}.csv'):
+        target_df = pd.read_csv(f'{save_path}/{camera_name}/{camera_name}.csv')
+        azi_target = target_df.loc[(target_df['azimuth'] == f'{azimuth}'), :]
+
+        target_name = azi_target['target_name'].tolist()
+        left_range = str_to_tuple(azi_target['left_range'].tolist())
+        right_range = str_to_tuple(azi_target['right_range'].tolist())
+        distance = azi_target['distance'].tolist()
+
+        return target_name, left_range, right_range, distance, azimuth
+
+    else:
+
+
+        return [], [], [], [], []
+
+    # if select_data == 'target_name':
+    #     return target_name
+    #
+    # elif select_data == 'left_range':
+    #     return left_range
+    #
+    # elif select_data == 'right_range':
+    #     return right_range
+    #
+    # elif select_data == 'distance':
+    #     return distance
 
 
 def str_to_tuple(before_list):
@@ -147,3 +190,15 @@ def str_to_tuple(before_list):
     tuple_list = [i.split(',') for i in before_list]
     tuple_list = [(int(i[0][1:]), int(i[1][:-1])) for i in tuple_list]
     return tuple_list
+
+
+if __name__ == '__main__':
+
+    front_target_name_W, front_left_range_W, front_right_range_W, front_distance_W, front_azimuth_W = \
+        target_info.get_target_from_azimuth(front_cap_name, 'W')
+
+    print(front_target_name_W)
+    print(front_left_range_W)
+    print(front_right_range_W)
+    print(front_distance_W)
+    print(front_azimuth_W)
